Clear pod cache on delete.

+++ pkg/registry/pod/rest.go
@@ -32,6 +32,7 @@ import (
 
 type PodStatusGetter interface {
 	GetPodStatus(namespace, name string) (*api.PodStatus, error)
+	ClearPodStatus(namespace, name string)
 }
 
 // REST implements the RESTStorage interface in terms of a PodRegistry.
@@ -77,6 +78,12 @@ func (rs *REST) Create(ctx api.Context, obj runtime.Object) (<-chan apiserver.RE
 
 func (rs *REST) Delete(ctx api.Context, id string) (<-chan apiserver.RESTResult, error) {
 	return apiserver.MakeAsync(func() (runtime.Object, error) {
+		namespace, found := api.NamespaceFrom(ctx)
+		if !found {
+			return &api.Status{Status: api.StatusFailure}, nil
+		}
+		rs.podCache.ClearPodStatus(namespace, id)
+
 		return &api.Status{Status: api.StatusSuccess}, rs.registry.DeletePod(ctx, id)
 	}), nil
 }
